Caesar shifter project
Put in a word/sentence you want to encode, and a shift number so it can be encoded. Then it gives you the secret word that only you can have access to based on you remembering the shift number. To decode it type in the encoded version and the shift number to get back the original word.